@@ -25,7 +25,7 @@ setup(
         "Operating System :: MacOS",
         "Operating System :: Microsoft :: Windows :: Windows 10",
         "Operating System :: Unix",
-        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
         "Topic :: Scientific/Engineering :: Artificial Intelligence",
         "Topic :: Scientific/Engineering :: Physics",
     ],
@@ -39,9 +39,9 @@ setup(
         "rich>=9.2.0",
         "numpy>=1.23.1",
         "scipy>=1.5.2",
-        "tensorflow>=2.4.1,<=2.14.1",
+        "tensorflow>=2.15.0",
         "tensorflow-estimator>=2.4.0",
         "tensorflow-probability>=0.12.1",
     ],
-    python_requires="~=3.9",
+    python_requires="~=3.10",
 )
