@@ -284,9 +284,9 @@ def tf_sgd(
 
     opt_sgd = tf.keras.optimizers.SGD(learning_rate=learning_rate, momentum=momentum)
 
-    for _ in range(iters):
-        step_count = opt_sgd.minimize(tf_fun, [var])
-        print(f"epoch {step_count.numpy()}: func_value: {tf_fun()}")
+    for ii in range(iters):
+        opt_sgd.minimize(tf_fun, [var])
+        print(f"iter {ii}: func_value: {tf_fun()}")
 
     result = OptimizeResult(x=var.numpy(), success=True)
     return result
@@ -524,7 +524,6 @@ def cmaes(x_init, fun=None, fun_grad=None, grad_lookup=None, options={}):
     es = cma.CMAEvolutionStrategy(x_init, spread, settings)
     iter = 0
     while not es.stop():
-
         if shrunk_check:
             sigmas.append(es.sigma)
             if iter > sigma_conv:
@@ -627,7 +626,6 @@ def gcmaes(x_init, fun=None, fun_grad=None, grad_lookup=None, options={}):
     es = cma.CMAEvolutionStrategy(x_init, spread, settings)
     iter = 0
     while not es.stop():
-
         if shrinked_check:
             sigmas.append(es.sigma)
             if iter > sigma_conv:
